DAY1/ChatModels/GeminiChatModel.py
from langchain_google_genai import ChatGoogleGenerativeAI
from dotenv import load_dotenv

load_dotenv()

model = ChatGoogleGenerativeAI(model="gemini-3-pro-preview", temperature=0)

result = model.invoke("What is the capital of India?")

print(result.content)